+++ golem-worker-service-base/src/parser/expr/mod.rs
@@ -3,4 +3,5 @@ mod let_statement;
 mod record;
 mod sequence;
 mod tuple;
+mod flags;
 mod util;
